move fuzz files into separate directory
